// types/alpha-vantage.ts
export interface Earnings {
  symbol: string;
  annualEarnings: AnnualEarning[];
  quarterlyEarnings: QuarterlyEarning[];
}

interface AnnualEarning {
  fiscalDateEnding: string;
  reportedEPS: string;
}

interface QuarterlyEarning {
  fiscalDateEnding: string;
  reportedDate: string;
  reportedEPS: string;
  estimatedEPS: string;
  surprise: string;
  surprisePercentage: string;
}